# api/src/main/resources/swagger-ui-3.0.6/swagger.yaml
---
swagger: "2.0"
info:
  description: "Simple phonebook aplication"
  version: "v1"
  title: "Phonebook Service"
  contact:
    name: "Oleksandr Butyter"
    email: "obutyter@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:9000"
basePath: "/"
tags:
- name: "Contacts"
schemes:
- "http"
paths:
  /contact:
    get:
      tags:
      - "Contacts"
      summary: "Get all contacts"
      description: ""
      operationId: "getAllContacts"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Contacts found"
        404:
          description: "Contacts not found"
    post:
      tags:
      - "Contacts"
      summary: "Create new contact"
      description: ""
      operationId: "createContact"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New contact"
        required: true
        schema:
          $ref: "#/definitions/ContactDTO"
      responses:
        201:
          description: "Contact created"
  /contact/{id}:
    get:
      tags:
      - "Contacts"
      summary: "Get contact by id"
      description: ""
      operationId: "getContactById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Contact found"
    put:
      tags:
      - "Contacts"
      summary: "Update contact"
      description: ""
      operationId: "updateContact"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "body"
        description: "Edited contact"
        required: true
        schema:
          $ref: "#/definitions/ContactDTO"
      responses:
        200:
          description: "Contact updated"
    delete:
      tags:
      - "Contacts"
      summary: "Delete contact"
      description: ""
      operationId: "deleteContact"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        204:
          description: "Contact deleted"
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
  ContactDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      firstName:
        type: "string"
      lastName:
        type: "string"
      phone:
        type: "string"